+++ tests/stdlib/tjsonmacro.nim
@@ -639,18 +639,25 @@ static:
 
 import strtabs
 
-proc testCustom()=
-  var t = {"name": "John", "city": "Monaco"}.newStringTable
-  let s = toJson(t)
-  doAssert $s == """{"mode":"modeCaseSensitive","table":{"city":"Monaco","name":"John"}}""", $s
+# xxx move to `tjson` pending https://github.com/nim-lang/Nim/pull/14572
+proc testRoundtrip[T](t: T, expected: string) =
+  let j = t.toJson
+  doAssert $j == expected, $j
+  doAssert j.jsonTo(T).toJson == j
 
-proc testToJson() =
-  var t = {"z": "Z", "y": "Y"}.newStringTable
-  type A = ref object
-    a1: string
-  let a = (1.1, "fo", 'x', @[10,11], [true, false], [t,newStringTable()], (foo: 0.5'f32, bar: A(a1: "abc"), bar2: A.default))
-  let j = a.toJson
-  doAssert $j == """[1.1,"fo",120,[10,11],[true,false],[{"mode":"modeCaseSensitive","table":{"y":"Y","z":"Z"}},{"mode":"modeCaseSensitive","table":{}}],{"foo":0.5,"bar":{"a1":"abc"},"bar2":null}]"""
-
-testCustom()
-testToJson()
+block: # toJson, jsonTo
+  type Foo = distinct float
+  testRoundtrip('x', """120""")
+  when not defined(js):
+    testRoundtrip(cast[pointer](12345)): """12345"""
+  testRoundtrip(Foo(1.5)): """1.5"""
+  testRoundtrip({"z": "Z", "y": "Y"}.toOrderedTable): """{"z":"Z","y":"Y"}"""
+  testRoundtrip({"z": (f1: 'f'), }.toTable): """{"z":{"f1":102}}"""
+  testRoundtrip({"name": "John", "city": "Monaco"}.newStringTable): """{"mode":"modeCaseSensitive","table":{"city":"Monaco","name":"John"}}"""
+  block: # complex example
+    let t = {"z": "Z", "y": "Y"}.newStringTable
+    type A = ref object
+      a1: string
+    let a = (1.1, "fo", 'x', @[10,11], [true, false], [t,newStringTable()], [0'u8,3'u8], (foo: 0.5'f32, bar: A(a1: "abc"), bar2: A.default))
+    testRoundtrip(a):
+      """[1.1,"fo",120,[10,11],[true,false],[{"mode":"modeCaseSensitive","table":{"y":"Y","z":"Z"}},{"mode":"modeCaseSensitive","table":{}}],[0,3],{"foo":0.5,"bar":{"a1":"abc"},"bar2":null}]"""